requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
Werkzeug==3.0.1
WTForms==3.1.1
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0
openai==1.12.0
